♻️ refactored search input into a reusable class
